Rainbow now working.

+++ src/main/scala/hlserver/effect/effects/Rainbow.scala
@@ -2,10 +2,10 @@ package hlserver.effect.effects
 
 import hlserver.effect.Effect
 import hlserver.util.{Color, ColorHSV, Strip}
-import org.joda.time.DateTime
+import com.github.nscala_time.time.Imports._
 
 class Rainbow extends Effect {
   override def render(index: Int, strip: Strip, time: DateTime): Color = {
-    ColorHSV(index / strip.length.toFloat, 1f, 1f)
+    ColorHSV(index.toFloat / strip.length.toFloat, 1f, 1f)
   }
 }
